[#36] 로그인 resolver 테스트 추가

+++ app/src/main/kotlin/com/santaclose/app/config/JWTConfig.kt
@@ -16,7 +16,7 @@ data class JWTConfig(
     val secret: String,
 
     @field:Positive
-    val expiredDays: Int,
+    val expiredDays: Long,
 ) {
     val key: SecretKey = Keys.hmacShaKeyFor(secret.toByteArray())
 }
